+++ eslint.config.js
@@ -24,12 +24,7 @@ const baseConfig = [
     languageOptions: {
       parserOptions: {
         projectService: {
-          allowDefaultProject: [
-            '*.mjs',
-            '*.js',
-            'public/*.js',
-            'scripts/*.mjs',
-          ],
+          allowDefaultProject: ['*.mjs', '*.js', 'public/*.js'],
         },
       },
     },